ft-printf-main-part

+++ ft_printf.h
@@ -19,14 +19,14 @@
 
 int		ft_printf(const char *s, ...);
 
-int     print_part(const char **ps);
+int		print_part(const char **ps);
 void	print_number(char *s, unsigned long n, unsigned long mode);
 int		nbrlen(unsigned long n, int mode);
 
 int		print_char(char c);
 int		print_string(char *s);
 int		print_pointer(unsigned long int n);
-int		print_decimal(int n);
+int		print_decimal(long n);
 int		print_unsigned_decimal(unsigned int n);
 int		print_hexadecimal_lowercase(unsigned int n);
 int		print_hexadecimal_uppercase(unsigned int n);
